Fixed the restoreValue() (the function in dDataControlMixin that restores a previous value of the control from the user settings file) so that if there was no previous value, the string "None" isn't filled in instead.
Removed the code in dDataNavSelectPage that clears the target control's
value if '-ignore-' is chosen. This goes back to conversations we had
a couple months ago regarding the ability for a user to select the 
dropdown value completely independently from the value in the target
control. If the user chooses '-ignore-' the value in the target control
will be ignored in the requery, but let the user erase that value or 
keep that value for convenience according to their own needs.

Added code in dDataNavSelectPage to save/restore the values in the 
target control to the userSettings.ini file. Next, I'll want to make
dDropDownList savvy to the userSettings table so that those settings
get saved/restored as well.