Made the limit and select options dropdown boxes on the Select Page save/restore their state between sessions. Unfortunately, due to limitations in the current implementation of the config file format, I had to change '  -ignore- ' to '-ignore-'. Didn't someone volunteer to make the .userSettings.ini into a .userSettings.xml? <hint hint> :)
Added a Choices property to dDropdownList that lets the options be
edited/queried at runtime, although it is pretty lame because you
can't resize the list that makes up the options at runtime. IOW,
you have to send the choices list to the constructor, and then you
are stuck with the choices list always being the same length. This
is a limitation of the wx.Choice control so we will need to either
implement our own dropdown list control or move to a different wx
control.