Modified the getBizobj() method of the form so that if the specified dataSource is "form", the form returns a reference to itself. This is useful for forms where you want to easily reference the controls and their values, but don't have the need for a database connection and a bizobj. It's also analagous to VFP's use of 'Thisform.property' to bind a control to a form's property.
Modified the getEventData() method of the uiwx module so that it extracts more of the native wx properties into the dEvents object's EventData dict. The properties have the same name as the methods in the wx event, but begin with a lower-case letter, as is the Dabo convention. Example: the native wx.EVT_COMBOBOX has a GetString() method that returns the text of the selected item. In the corresponding dEvents.Hit object, you can now reference 'evt.getString' to get the same information.