Added 'dGridSizer', which is based on the wx.GridBagSizer.
If you specify a particular cell when adding items to the sizer, they will be added into that cell, but if no position is specified, the item will be added to the first available cell. The sizer allows you to specify either the MaxRows or MaxCols; it will add items until that max is reached and then begin a new row/column. The default is 2 columns.

Using this, you can populate a rectangular array of items without having to manually calculate the position of each. For example, if you have a series of labeled controls, you can append a label, then a control, then another label, then another control, etc., until you're done. With the default of 2 columns, you will get the items arranged with all the labels in the left column, and all the controls in the right.

Still to be added: control for making columns and rows "growable", since by default the dimensions are fixed.