Got rid of the dApp property AutoNegotiateUniqueNames, and instead made Dabo smarter about when to make a name unique by appending a number like 'dTextBox1' and when to raise an exception. Basically, whenever the user explicitly sets the Name, and another sibling already has that name, a NameError will be raised. Whenever Dabo is just trying to use a default name, a number will be appended and a NameError will not be raised.
By the way, it turns out that setting AutoNegotiateUniqueNames to False before
would have resulted in no objects being able to be created (the NameError was
always raised).

I had to touch pretty much all class defs in ui/uiwx, because that is where the
object gets its default name.