Added a smaller rendition of the dabo with lettering bitmap.
Added properties to specify whether a status bar is automatically
created or not, and what if any menu bar to use.

Started work on a new Help|About window. I intended to be able to 
put this in the ui-agnostic dabo/ui directory but there are still
too many areas where we need to use wx directly, unfortunately. I'll
see about reworking it to not use DC's to do the drawing, and instead
use dabo objects. That should let it be ui-agnostic.