Cosmetic: changed the display of '-ignore-' dropdowns and their fields to italic when they are in the '-ignore-' state.
Moved the logic for sizers drawing their outlines to the sizer classes, so that it can be used outside of the data nav forms. Left some hack code for the dGridSizer's outlines in the data nav form due to a bug discovered in the wx sizers.

Added a menu item to the data nav form to turn the display of sizers on/off. Changed the data nav pages so that they get their 'draw sizer' status from the form instead of storing that individually.