The comparison operation dropdowns in the datanav select page were not having their values saved in the .userSettings.ini file, because that uses the control's _oldVal property, not its Value, and the initial settings are set programmatically. This resulted in a series of errors being throw when a datanav form is run for the second time, since the .userSettings file recorded an empty string for the Value, and restoring it failed, since their is no choice in the dropdown for an empty string.
I changed the dDataControlMixinBase.refresh() to update _oldVal whether or not the control is data bound, and added a call to dropdown.refresh() after the initial value is set.

This brings up an issue: if we are going to be saving/restoring the status of controls, and not all of them are going to be bound to data, any that we set programatically will have the same problem. Two questions: first, in the dDataControlMixinBase.saveValue() method, there is the following comment:
		# It is too late to get Value directly:		
		value = self._oldVal	
Why is it "too late"?

Second, can we tie the refreshing of self._oldVal into the Value setting methods? This way any update, either programmatic or user-initiated, will result in that prop being updated.