Ok, fixed a bozo bug I introduced: Target could be None.
Also, reviewed the saveValue() and restoreValue() code, and determined 
that there is nothing in there to prevent value from being saved/restored
whether or not we are connected to a database. It is all controlled by
the setting of the SaveRestoreValue property. I believe this could cause
problems if you had SaveRestoreValue set to True, and the control is
bound to a data field in a table, but haven't tested that. It should work
in the case where DataSource is "form", but I haven't tested that either.