Modified dDropdownList to be able to have its Value property bound to a key, which could be a key in a database table, for instance.
Please run the test: python dDropdownList.py

...and review the changes. I'd like to make similar changes to dRadioGroup and other
controls that should get similar functionality.

Theoretically (I haven't tried it yet), if the ValueMode is 'key', and the Keys dict
has been properly set up, then if the DataSource and DataField are set, the bizobj
should get the proper changes (the key, not the position or string value).