Fixed the print preview so it works again, ugly though it is.
Added a hook for user code to override the control class for the edit page. I
used this in one of my apps already, to display a dropdown of client names 
keyed on iclientid, instead of showing the iclientid in a textbox. (This also
confirms that the dropdown changes work against real data, which was only
theoretical before).

Set the label alignment to top for the labels on the edit page, because middle
looks wrong for edit boxes that grow.
 
Hid some key errors in dDropDownList: hopefully I don't regret it! :)