Created a Dabo subclass of the wx.SplitterWindow, and a subclass of dForm that incorporates the splitter window and its main properties.
Also revised test.py so that it allows for the testing of classes based on wx.Frame, since the old version would only work for controls that could be added to the wx.Frame it created. While I was mucking around with this program, I changed the sizing of the frame, as this was usually coming up way too small in most of my tests.