This is a large commit, as my scope kept creeping as I went: In the course of changing something that needed changing, I'd find something else that needed changing, etc.
---> The demo's, the appWizard, and various files in daboide need
---> to be updated as well. Those commits will follow this one.
---> Applications developed from older versions of the wizard should
---> be fine, though.

Refactored the boilerplate __init__ code from the individual classes
in ui/uiwx to dPemMixin. Now, the individual __init__ code is short
and unique to each class. Every __init__ now has this signature:

	def __init__(self, parent, properties=None, *args, **kwargs):

Made a concerted effort to remove all afterInit() and initEvents() 
from the framework code, replacing with _afterInit() and _initEvents()
so as to not conflict with user overrides.

There are certain properties that can't be set on an object after the
wx object has been fully constructed. I've refactored how I account
for this, which is already working better but there may be some problems
still, and I have some ideas on how to build on what I've started to
eventually get rid of the need for the user to remember which properties
need to be set before the init phase.

dListBox, dDropdownList, dComboBox, and dRadioGroup all share some common
properties. Last week, I made some changes to dDropdownList to allow for
setting a ValueMode which can be "position", "string", or "key". I've 
extended this to the other controls. Spent time getting dRadioGroup,
dComboBox, and dListBox working well with abstractions that make sense.

Reviewed and modified some of the property docstrings in some of the 
classes. This is something that needs to happen in an ongoing fashion.

The wx arguments to the constructor, while they still get passed to the 
wx object, are no longer interpreted by Dabo. So, the 'name' argument
is now 'Name' as that matches the property. All addObject()'s must now 
look like:

	self.addObject(dLabel, "lblTest")
	self.addObject(dLabel, Name="lblTest")
	self.addObject(dLabel)
	and not:
	self.addObject(dLabel, name="lblTest")
	
This was to make addObject's arg style match the arg style of the object
constructors.