Added intelligence in dTextBox to be able to convert datetime.date and datetime.datetime to string values and back again. The string format is ISO 8601 ("2004-12-16" and "2004-12-16 13:29:33") and as it is written it is very strict on that formatting. If the user strays from that formatting it won't work.
This is just the raw low-level behavior that we are defining here.
dDateTextBox will be able to build on this base behavior to allow
multiple date/datetime display formats and to be friendlier to
the user.

Ed, this won't work with your specific problem because it looks like
your MySQLdb is bringing in your datetimes in mx.DateTime format
which I didn't code for - I only coded for Python datetime.datetime.
But run the test in dTextBox.py and see if at least that part works
as expected.

Altered test.py to make the controls wider by default.