Found out that MySQLdb was using the mx DateTime classes if they happened to be installed on the machine on which Dabo was running. So I check for this in the cursor, and if mx DateTimes are found, they are converted to native datetime.datetime values.
I also have some code I added to the textbox to allow it to handle mx.DateTime values. I was going to remove it, since the cursor removes these, but thought that in those cases where Dabo is used without cursors, it would be nice to be able to handle all sorts of dates.