The chain of code fired for every dEvent instantiation has two significant performance bottlenecks: logging the event, and setting the native event information. Therefore, I've made two global settings, defaulting to False, to determine whether these two blocks of code run.
Event logging is a no-brainer: we simply don't care about this at
runtime, so it is set to False by default now. The designer or a 
user can turn it on manually if they need to, by simply issuing:

	dabo.eventLogging = True
	
...and when they are done, setting it back to False.

The code that was recently added to insert all native event
information into the EventData property is the biggest bottleneck,
though, as it runs an eval() on every attribute of the native 
event object and stuffs the result of the eval into the EventData
dictionary. This is very useful, but IMO we only care about it 
during development, when determining what useful things to pull 
out of the native event explicitly. So this is turned off by default
now as well. Turn it on at any time:

	dabo.allNativeEventInfo = True
	
... and remember to set it back to False when you are done.

Also went through dEvent and optimized where I could.