Reverted my hackish changes to the fieldSpecs (the <join> and <expression>), and modified the appWizard to define the sql in the bizobj instead of by using the fieldspecs in the UI.
Added a new property NameBase to dPemMixin which assigns a Name to
an object but has Dabo find the next unique name by adding integers
after it. This is basically how the Name property used to work. 
Usually, user code will want to assign to the Name property because
if an object already has that name it is an exception to handle. 
However, sometimes you don't care and just want to get a unique name.
Changed the generated code in appWizard to name forms using this 
NameBase property, so that the end user can instantiate more than 
one instance of the same form without an exception being thrown.

Also see my imminent commits to daboide and dabodemo.