Added a class property named '_IsContainer' to all the base classes. This is used by the dPemMixin.__getattr__() method; if self._IsContainer is False, it will not attempt to call the FindWindowByName() method of the object, since there is no way that a button or a line can contain child windows.
This change was made because of a bug that took me quite a while to track down. Occasionally an app with a custom menu would just hang when I tried to quit out; I eventually tracked it down to an infinite loop that was caused when __getattr__ would try to call the menu's FindWindowByName() method, which it doesn't have, as it is not based upon wx.Window. This would cause __getattr__ to be called, which resulted in the infinite loop.