Commented out the code in dPemMixin that set the Size of any object to (0,0) if it was the default of (-1, -1). It was interfering with the auto-sizing of panels when their Layout() method was called. I tested this change with multiple programs, and it didn't seem to break anything.
Added a debugging routine 'listMembers()' to the SizerMixin. This will list all of the children and their sizer settings, and optionally recurse down all child sizers. Also added code to dGridSizer to enable it to also use the Orientation property.

The biggest changes are to dDialog, though. Pretty much re-worked the entire class, completely Dabo-izing it by adding dPemMixin and re-factoring the creation methods. Also added an AutoSize and Centered properties to control the dialog's size and position when it is shown.