Added the following properties: Border, BorderBottom, BorderLeft, BorderRight, BorderTop, and Spacing. They reduce the amount of repetitive code when laying out a form as they allow for default settings.
If Spacing is set to a positive amount, a spacer equal in size to that amount is inserted between items. It uses the new 'addSpacer()' method that accepts a single integer and creates the appropriate spacer based on the sizer's Orientation. 

If Border is set to a positive amount, and no 'border' parameter is passed, this value will be used for the border.

If no 'borderFlags' parameter is passed, and any of the BorderXX properties are set to True, those sides will be set as the flags. If none of the BorderXX properties are True, the original default of "all" is used.

Example of usage:

sz = myForm.Sizer
sz.Spacing = 10
sz.Border = 25
sz.BorderLeft = sz.BorderRight = True
...
sz.append(itm)
# This will add itm to the sizer, with a 25-px border on 
# its left and right sides
sz.append(itm2)
# This will first insert a 10-px spacer, and then add itm to the sizer, 
# with a 25-px border on its left and right sides
sz.addSpacer(44)
# This will add a 44-px spacer. If the sizer is vertical, 
# the actual spacer inserted would be (1, 44)
sz.insert(0, itm3, border=5)
# This will insert itm3 at the beginning of the sizer, with a 5-px 
# border on the left and right sides. A 10-px spacer will be
# added after it
sz.BorderLeft = sz.BorderRight = False
sz.append(itm4)
# This will append the 10-px spacer and then itm4. It will have
# the 25-px border, but this time, since all the BorderXX flags
# are False, it will be added to 'all'.