Updated the definition of daboDesignerClasses to include all the current classes.
Fixed a problem with dPemMixin. Previously, once a control's size had been changed through either the Size, Width or Height props, you couldn't change the size to anything smaller. The use of SetBestFittingSize() instead of SetMinSize() allows for full control over resizing, and also updates the sizers so that the control is correctly positioned.