Added the dUICursors class to wrap the wx.Cursors class. I specified the name as including "UI" to make it clear that these did not have anything to do with the data cursor classes.
Reinstated the call to getColorTupleFromName() in the PemMixin code, as it turns out that wx only implements a small percentage of the color names.

Fixed a bug in the sizer code that did not pass along any proportion when a spacer was added. Also made the method names for spacers consistent with the overall append/insert pattern.