This makes the SpinnerWrap property work when sent to the constructor. Otherwise, the property had to be set in the subclass beforeInit() hook.
I'm currently working out the best LCD for handling all properties
in a generic fashion, so the user doesn't have to know which props
to put in beforeInit() (or initStyleProperties()) and which to put
in afterInit() (or initProperties()). 

By maintaining a list of the props that must be set prior to init,
at least the class knows which ones to preinit and which to postinit,
which it can do for props passed to the constructor. For the case of
properties in code such as the preinit and postinit hooks, at a 
bare minimum at least this same list could be consulted and a warning
issued if user code tries to set a property in the wrong place. 
Stay tuned...