Reverted the previous change of combining the dCommandButton and the dBitmapButton into a single class based on dBitmapButton, due to display incompatibilities. dCommandButton is now depracated, replaced by dButton. There is no change in functionality, just a simplified name.
Also updated dDataNavPage to use dButton instances, and fixed (well, hacked something that works, for now) a grid display problem on the Browse page. Before this the sizer for that page wouldn't recognize the size of the grid until the form was resized, so I added a one-time resize that forces the sizer to lay out the page properly.