Fixed a bug in dBizobj that only showed up in Bubblet. The bizobj was checking for changes upon closing, and since there are no cursors, it was throwing an error.
I thought I updated the dApp/uiApp stuff last week, as I'm using it in the latest Designer. It simply gives you some control over when the main form of the app is displayed. In the case of the Designer, if they hadn't already specified a class to work with, I wanted to open up the class selector without having the main form open yet, since the form to be used will depend on the selected class.