Removed the unused escapeQt function from dObject.
Implemented function dObject.getValidEvents() that returns a list of dEvents
that apply to the object. The logic for determining whether a given event applies
is encapsulated in dabo.dEvents, in each event's appliesToClass() classmethod.

Hopefully this will be useful in the designer, to let the user define their
event bindings.