Added a method 'callAfter()' in uiwx. It basically wraps wx.CallAfter so that this useful technique can be used in code written in Dabo.
Added removeRow() and removeCol() to the dGridSizer class. This will delete the items in the specified row/col, and move the higher rows/cols over to fill in the void.

so that --This line, and those below, will be ignored--

M    ui/uiwx/dPageFrame.py
M    ui/uiwx/__init__.py
M    ui/uiwx/dGridSizer.py