Further work on menus. The wx menu events are confusing and inconsistent, depending on whether the menu is a child of the menu bar or another menu, so I consolidated the events. Now, there are only two menu events: Hit and MenuHighlight. MenuHighlight gets raised whenever a menu or menu item gets highlighted, whether it is a top-level menu (File, Edit, etc.) or a submenu, or a normal menu item. Removed the MenuClose event which never fired in my experience anyway, as well as the MenuOpen event which only fired for top-level menus and which I consolidated into MenuHighlight.
Also, it turns out that we don't need to set the Form reference in the 
menubar - there are methods added by wxPython (not wxWidgets) that keep
track of that for us. You can change a form/menubar relationship in one
of two ways:

form.MenuBar = mb
mb.Form = form

TODO on menus: Enabled property, checkitems, radioitems, Value property for
checkitems and radioitems, Children property for menus, being able to
reference menubar.filemenu.openmenu.menuitem (chain of names like we can do
for UI containers).

Possible TODO: DynamicEnabled, which runs the specified function to determine
whether to enable the item/submenu. Would run automatically when the menu was
opened. Kind of like VFP's skipfor but it wouldn't be a character expression
but a first-class function. A reference to the menu/menu item would be passed
to the specified function. Does that sound useful or no? Perhaps it isn't 
necessary or desired in a oop menu after all.