Finally completed the separation of the Dabo cursor from the bizobj, allowing it to be used directly. I had had these changes sitting around for a while, but never finished testing them.
The dConnection class is greatly simplified; it will now accept a connectionInfo dictionary, or even a path to a .cnxml file, and will be able to create the connection. There is no longer a need for a separate step of creating the dConnectInfo object first if you already have a dict or .cnxml file.

To use a raw cursor, this is all you need to do:

>>> import dabo
>>> conn=dabo.db.dConnection("sample.cnxml")
>>> crs = conn.getDaboCursor()
>>> crs.execute("select czip from zipcodes limit 10")
10L

Note that you still have to set the KeyField, Table, and AutoPopulatePK properties in order to make it into a fully-functional cursor. 

There is now a new method named 'executeSQL()' designed to work with the query builder functions. After you do all the setWhere(), addField(), etc., stuff, you still had to do: crs.setSQL(crs.getSQL()). Now you can simply call executeSQL(), which will create the SQL from the query builder props, and pass it to execute for you.