x, y, width, and height can now be specified as numbers or strings in the spec file. Numbers get interpreted as pt values, strings get interpreted based on the unit specified ("3 in", "10.4 mm", "3.2 pica", etc.).
Added a report dict which will be available to the running report to provide
information for use by user expressions as specified in the spec properties.

Added vAnchor and hAnchor properties, which specify that the y and x values
refer to "top", "center", or "bottom" and "left", "center", or "right",
respectively.

Made use of anchors and the report dict in the positioning of the Dabo icon: 
try changing the page width and notice that the dabo icon always stays 
anchored to the right side of the page footer - because of the dynamic expr
in the x property for the image.

Page size and orientation can now be specified like "letter", "legal", or "a4" 
and "portrait" or "landscape", respectively.

Note that while I'm designing the report writer and spec file with a high 
amount of runtime flexibility and dynamism, the report designer will 
necessarily only write spec files with a subset of the features, at least 
at first.