Integrated the option of force closing a form by passing an optional boolean 'force' parameter to the form's close() method. Now calling form.close(True) will not call the beforeClose() code, leaving no chance of preventing the form from closing.
Reworked the naming of the methods involved in the closing of the form. Assuming that force=False, the __onClose event handler will still call _beforeClose(), which calls beforeClose(). If the form is going to close, the method called for executing any cleanup code has been renamed from 'close()' to 'closing()'. A new method named 'close()' has been added to wrap the wx method 'Close()', and will accept an optional 'force' parameter.