Added show() and hide() to the form class.
The wizards are now working. There is a sample of some pages in the test section at the bottom of dWizard.py; you can see them in action by running the dWizard.py script directly. The methods should be fairly self-explanatory. The example classes illustrate conditional behavior, such as skipping pages and not being able to move in a given direction until conditions are met.