Refactored reportWriter a bit.
Began a crude report designer. So far it instantiates a report writer instance,
loads in the samplespec.rfxml file, and correctly lays out the position and
size of the pageHeader and pageFooter bands. If you click on the band labels,
you can modify the band height, which is reflected visually on the designer
surface. That's it for now though.

The report writer instance will be used internally by the report designer for
generating previews and also for sharing the common conversion functions and 
report layout dict.

Cleaned up the font setters in dPemMixin, in an effort to get the font
properties to work correctly during instantiation. I was able to reduce the
problem I'm having all the way down to a simple wx-only script though so I 
passed the question on to wxPython-users.