Began implementing fileNew, fileSave, fileSaveAs, fileOpen, fileClose, most of which will cause errors. But you can feed specfiles on the command line, such as:
python reportDesigner.py samplespec.rfxml

...and while your changes won't be saved, at least the designer/writer know
how to go about saving your changes if only there were code in the right
places to ask them to do it. :) File|New works, but after doing that you need
to resize the form for the new report form to render on the design surface.

Next up: finish the file management skeleton. Add rendering of objects and
editing the properties of the objects. Add editing of page properties too.
Could be a few days because I've got to switch back to my day job now...