Fixed a bug in getCaptureBitmap() - this should fix those problems Ed was having originally with garbage-filled bitmaps.
Added visual indication that the report band is dragging by getting a 
capture bitmap and blitting it to the screen. At this point I don't 
understand how to erase the background so that the trails don't appear,
but am congratulating myself on figuring out how to at least get this far.