Added better encoding support to the dCursorMixin class.
Refactored the code to convert color names to color tuples from dPemMixin and into dColors so that it can be used by other classes that don't inherit from dPemMixin.

Dabo-ized a good deal of the ImageList code for the controls that support ImageLists. Modified the datanav code to use the new wx-less image code.

Added image support to dTreeView, as well as basic Font support. There seems to be an issue with the GetFont() function of tree items, which is preventing me from adding FontItalic, etc., support. I may look into this, but it is not a priority. Also added the concept of 'Descendents', which is all children, grandchildren, etc. of a node.