I'm about to ditch samplespec.py, since the xml way is working so well (yes, I was surprised, but xml is easier to deal with). Anyway, I wanted to get some of the other elements over from samplespec.py first, and I made the expressions work in the report designer (the x position of the image in the footer is derived from a runtime expression - this now works in the designer just like it works in the runtime report writer)..
Next up is a major refactoring of the report designer before this becomes
a house of cards. As it stands, the objects know too much about each other
and I want to make each object more of a black box. I've proved most of the
concepts, now I've got to commit to an overall design and set a firm foundation
for the "filling in the blanks".