Added dFontDialog. Still need to Dabo-ize the font objects themselves, though.
Added a 'release()' method to dPemMixin. Changed all the references to Destroy() to call release() instead.

Created a simple directory browsing method in the dTreeView control. A new method, named 'makeDirTree()', will accept a starting path and an optional wildcard pattern, and create a tree representation of that directory. It is fairly slow with large directories, and when a wildcard is specified, and no files are found in the directory that match, the directory itself is still added. But it does work!