Fixed a problem with reportWriter that wouldn't let the designer work with new specfiles.
Added appropriate mouse cursors depending on whether the object is selected and
whether the mouse is hovering over an anchor or not. Next step will be to allow
the user to move/resize the object from the keyboard and mouse. After that, I'll
need to figure out how to get the prop sheet integrated per our discussions last
week.