Added new class 'dImage' for generic image handling. Right now it will instantiate with no image, or with an image path passed to the constructor with the 'img' parameter. This parameter name will probably change; I was thinking of naming it like VFP's 'Picture' property - any opinions?
Tightened up the parameter names for the various dialogs. Abstracted out the process of selecting files and folders so that you no longer have to instantiate the dialog, check its return value, and grab the selected path. Now you can call 'dabo.ui.getFile()' (or getFolder, or getSaveAs) directly, and the return value is either the selected item, or None if the user canceled out.

While I was at it, I moved getMouseObject() into __init__.py.