Now, you can click on an object to select it, and then use the arrow keys to move it around, and shift-arrow keys to resize it.
Added a required "name" property to objects - the designer will need to enforce
unique names and fail to load a rfxml that doesn't have unique names, but it doesn't
do that yet.

In other news, I was able to execute 'pdftoppm test.pdf test' on my linux box,
which converted my 3-page pdf file into 3 separate ppm files as test-00001.ppm,
test-00002.ppm, and test-00003.ppm. I was then able to use the dImage.py demo
to display these ppm files on screen. Something like this can be used in the 
future to 'print preview' the dabo reports within a dabo app. The problem will
be, finding a way to get pdf->image in a cross-platform way (pdftoppm is part
of the xpdf package, is GPL, and only built for Linux as far as I can tell).
I was hoping the Python Image Library would help, but no such luck.