Modified the binding of controls. A control may now be bound to the form or to object that can be referenced as 'self.X', with 'X' being any resolvable expression. Thus, a DataSource of 'form' will bind the control to a form's attributes, while 'self.Parent' will bind the control to its Parent container's atts.
Also, the DataField can now also reference a method of its bound object. In that case, the value in the control will be read from the object on refresh(), but changes to the control's value will not be written back, as this would overwrite the method itself.

I've tested this with all the apps on my machine; please test with yours and let me know of any problems.