Added the import of the logPoint function to help debugging.
Created two new methods in dColors: randomColor() and tupleToHex(). The former returns a random color (surprise!), and the latter converts something like (46, 139, 87) to #2E8B57.


Modified sizers to accept keyword arguments in their constructors. Added method 'copyGrid()' to the grid sizer that will move the contents of one grid sizer into another. Right now I plan on using it in the Designer, but it may have uses elsewhere, so I felt it was right to put the code up in the base class.

Corrected bug in the form mixin's layout() method that prevented forms from updating themselves in some cases.

Updated the dListControl to work better with MultiSelect.