Changed the dynamic import in dBackend to not use 'exec()'
Added SortingForm to the dialogs directory. This accepts a list and displays a visual means for sorting. Also added a wrapper function sortList() to uiwx; this accepts a list as a parameter, and returns the sorted list if the user clicks 'OK', and the original list if they cancel out.

Modified the getfile dialog code to recognize typical extensions and create the wildcard string accordingly.

Added some safety code to the dControlItemMixin class for the event where an index is passed that is outside the range of the list index.