Refactored out the sorting controls into a ListSorter panel containing the controls used to sort the lists, and a pre-built form that uses this panel. This enables the controls to be dropped onto any form.
Changed the WizardPage to inherit from dScrollPanel, so that longer pages will still be accessible.

Changed all of the 'leaving' checks in the Wizard to only prevent the page from changing if False is returned. Previously, forgetting to explicitly return True prevented the page from changing.

Added a layout() method to dScrollPanel to bring it in line with the other panels.