Reimplemented new and delete for child records in datanav. They are still pretty flaky though: I can't seem to add new records to the child yet - it looks like it works but then I refresh and there aren't any child records.
I was getting annoyed by a side-effect of a recent change to 
dForm: a messagebox whenever I navigated to BOF or EOF. Fixed
this back to just adding "BOF" or "EOF" to the status bar.

While I was in dForm, I added the name of the datasource 
to the default delete messagebox.