This fixes the datanav-with-child problem where the focused control's value wouldn't update when navigating. This actually brings up a gotcha:
wx.Window.SetFocus() doesn't raise wx.EVT_KILLFOCUS on the control
that previously had the focus.

We need to put on our todo list a setFocus() that manually raises
dEvents.LostFocus on the previously selected control before calling
SetFocus to the new control. And then we need to stop using
SetFocus().