EXPERIMENTAL!
Rather than work on dGrid, which is being used directly, and which will prevent Dabo from running if left in an unstable state, I've created a separate file named 'dGridX', which will not be integrated into Dabo until all the kinks are worked out. I'm checking this in now so that a) it will be backed up with all the rest of the project stuff and b) so that Paul and I can discuss what I'm working on.