Added property Tag, which user code can use for specific purposes. While I never used this property in VFP, I do appreciate its purpose and came up with a use-case just now for specific menu items in the datanav framework (see below).
Enhanced dMenu.append(), insert(), and prepend() to take additional
**kwargs to pass on to the dMenuItem constructor.

Enhanced the datanav Navigation menu to include all edit pages,
not just the first edit page. This takes advantage of the Tag
property in dMenuItem to save the datasource name for use by the 
generic onEditCurrentRecord() event handler to determine which
edit page to bring forward.