I'm doing some testing with using Dabo in conjunction with a mod_python web application, and I needed to be able to load my bizobjs and db connections as defined in the cnxml file, but in order to get that loaded easily I needed the dApp object and its _initDB method. To get that I needed to call app.setup() which also inits the UI. So for now I added a initUI=True argument to dApp.setup() so I can continue testing this. Eventually, I think we'll be able to have a uiType of mod_python and be able to instantiate dApp normally.
Along these same lines, I had to make a change to the loading
of the cnxml file, as it wasn't finding the file unless I
gave it an absolute path, which dApp already provided as
self.HomeDirectory. The change I made doesn't appear to 
have any bad effects when running Dabo normally.

Wow, that is a verbose changelog for just a couple insignificant
changes. I'm coming down with something and my brain is kinda
loopy. Bear with me.