Re-ordered the call to beginTransaction() in the bizobj so that it comes after the initial condition testing.
Added the 'browse' command to uiwx/__init__.py. It works pretty much like a VFP Browse command. Pass it a data source, and it displays a separate window with a grid in it containing the data. The data source can be a record set, or it can be a Dabo cursor, or any object with a getDataSet() method that returns a Dabo record set.
In order to do this, I had to also import the dGridX module, but all my testing shows it to be stable and doesn't mess anything else up.

Changes to the dGridX:
- defined default renderers and editors for cells.
- eliminated several places where multiple calls to refill the table were being made.
- Added some basic unicode encoding into the grid. However, I've still run into some cases where non-ASCII text throws an error. This will be investigated further.
- Refactored the MoveColumn logic into the grid instead of the table.
- Dabo cursor temp fields are now excluded from the grid when filling from a record set.
- Added the 'autoSizeCol()' method. This a column number and auto-sizes it to fit the data and the header caption. You can also pass "all" to this method, and it will auto-size all columns.