Big change for Firebird users!
If you have a generator assigned to a Before Insert trigger that returns the next value for your PK, Dabo will now pre-fetch that value before an insert and populate the PK field for you. This will enable it to be propagated to child records for their FK link to the parent table.

dCursorMixin calls pregenPK() before an insert of a new record. For backends that don't require pre-generated PKs, this call does nothing.

dBackend now has a method named 'pregenPK' that by default returns None. It is overridden in dbFirebird to determine the applicable generator and grab the next available value.

I've also re-instituted the flush() behavior of calling commit(). The original commit is firing, but it does seem that the second commit is required sometimes. In any case, it doesn't seem to hurt.